ARC-1945: Add API key authentication to cortex (#22)
Added a basic authentication mechanism that lets you use a fixed API key for all requests. Simply set the `API_KEY` environment variable on the server to force the client to specify the same API key in all requests via the `api_key` header or the `api_key` querystring parameter.