Preserve HTML integrity in chunker
The chunker now detects whether it's working with HTML or not. If HTML, it uses HTML elements to determine chunk. This prevents the scenario where <p>foo bar</p> is split into two chunks: `<p> foo` and `bar </p>`.

The HTML may also contain text-only nodes. If that's the case, then the chunker detects that and splits the text according to our regular chunking rules. This helps the scenario where someone may have a large body of text between HTML elements.

If this chunking strategy results in a chunk that's larger than the chunk size, an error is thrown.