typeDef generator should detect string arrays as potential inputs
When an array is defined as an input to a pathway, the default assumption is that it is a message. For one of my custom pathways, I needed a string array.

So I ended up defining it as:
```
        keywords: ['']
```.

With this code update, the empty string is detected as a string and cortex builds the pathway with the correct GraphQL signature.