ARC-2142: Cortex HTML chunker does not handle long blobs of text within HTML
Since there is no way to split a HTML element effectively, the cortex chunker is designed to give up if it encounters a large HTML element.

Example: <p>(long paragraph that may contain other html as well)</p>

However, per the current implementation, it complains even if the large element is non-html.

Example:
<p>(short paragraph)</p>
(long text-only paragraph)
<p>(short paragraph)</p>

It should be OK to chunk the long text-only paragraph instead of erroring out.