Allow overriding pathway execution code without overriding entire resolver
The use case for this is the one when the pathway would like to execute arbitrary JavaScript code before or after the prompts are executed.

To use this feature, the pathway implements the `executeOverride` method (named to indicate that this will override pathway execution logic). The `executeOverride` method supports two arguments: `args`, `runAllPrompts`. `args` gives the pathway access to all the input args, while `runAllPrompts` can be used to execute all prompts that are defined in the pathway using all relevant configuration options such as chunking, chunk size, inputFormat, etc.