Add support for DALL-E 3 API
DALL-E 2 and DALL-E 3 APIs are not compatible. As well, there doesn't seem to be an async version of the DALL-E 3 API, which is the version we were using for DALL-E 2. So created a new model type for DALL-E 3 and implemented the execute method. To support async operations, we keep the request open and send progress updates to the client over a websocket.