Reload pathways on publish and remove
We were checking for new pathways on get, but were not checking on publish and remove. This prevents a user from publishing and unpublishing in quick succession when session affinity is turned off (e.g. the user's publish request would go to one instance and then the unpublish request would go to a separate instance and fail).