File handler should use container-scoped hash keys (#406)
* File handler should use container-scoped hash keys

Since the file handler supports multiple containers now, we need to scope the hash keys according to which container the file is being placed in. Otherwise, you end up with a situation where a hash matches a file in a container other than what the user intended.

For backwards compatibility, files being placed in the default container don't use a scoped key.